RoPE Triton kernel: FP64 -> FP32 internal math for GB10 throughput
The fused RoPE kernel ran all rotation math in tl.float64 to match the
complex128 freq storage. On Blackwell (GB10 / DGX Spark) FP64 is roughly
1/64 of FP32 throughput, vs ~1/2 on H100, so this hot path was paying a
significant penalty for precision the downstream bf16 store cannot
preserve anyway.

Switching the in-kernel math to FP32 (with a cast after load, leaving
the freq storage untouched) measures:
    generate per frame at steps_size=4: ~321ms -> ~302ms (-6%)
    end-to-end FPS: 2.74 -> 2.89 (+5.5%)

Numerical sanity (S=256, H=20, C=128 vs FP64 reference):
    relative L2 error: 9.95e-06
    max abs error    : 0.0078 (one bf16 quantum)
    output dtype     : torch.bfloat16 (unchanged)

Constraint: GB10 / sm_121, NGC nvcr.io/nvidia/pytorch:26.03-py3 container, Triton 3.6.0
Constraint: bf16 output dtype is unchanged; downstream code is unaware of the math precision change
Constraint: freq tensors remain stored as complex128 / viewed as float64; this avoids touching every caller of init_rope_cache
Rejected: Storing freqs as FP32 across the model | invasive, requires changing init_rope_cache in dynamic_model and tokenizer plus all call sites; do separately if the cast-inside-kernel approach proves to be a memory-bandwidth bottleneck
Rejected: Keeping FP64 math because RoPE rotations need precision | the bf16 output store would discard the FP64 -> FP32 difference (max abs error already at the bf16 quantum); FP32 here is precision-equivalent end to end
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: If you ever store freqs as FP32 directly, also drop the in-kernel `freq.to(tl.float32)` cast; loading FP32 from memory is what unlocks the full memory-bandwidth win
Directive: Re-benchmark after any Triton upgrade; FP32 throughput on GB10 may improve with newer compiler versions and shift this optimization's payoff
Tested: rope_apply_fused_kernel_triton vs rope_apply_ref on (B=1,S=256,H=20,C=128) input inside the NGC container; relative L2 9.95e-06
Tested: Live websocket render at steps_size=4 cold; CUDA graph captures cleanly; sweep_harness.py shows hold_w/idle_a=1.92, pan_r/idle_a=3.43 (responsive)
Not-tested: Multi-hour stability run; timing was sampled over ~3 minutes of synthetic action drive
Not-tested: Same patch under steps_size in {1,2}; expected to scale similarly but not measured this commit

+++ src/kernels/rope_triton.py
@@ -15,9 +15,12 @@ def _rope_apply_fused_kernel_triton(x_ptr, freq_ptr, out_ptr, S, BLOCK_M: tl.con
     offs_m = tl.arange(0, BLOCK_M) + idx_s * BLOCK_M
     offs_n_x = tl.arange(0, C)
     x = tl.load(x_ptr + offs_m[:, None] * H * C + offs_n_x[None, :], mask=offs_m[:, None] < S, other=0.0)
-    x = x.reshape(BLOCK_M, C // 2, 2).to(tl.float64)
+    # FP32 math: GB10 FP64 throughput is ~1/64 of FP32; keep freq stored as
+    # FP64 to avoid touching callers, but cast to FP32 before the multiply.
+    x = x.reshape(BLOCK_M, C // 2, 2).to(tl.float32)
     x0, x1 = tl.split(x)
     freq = tl.load(freq_ptr + offs_m[:, None] * C + offs_n_x[None, :], mask=offs_m[:, None] < S, other=0.0)
+    freq = freq.to(tl.float32)
     freq = freq.reshape(BLOCK_M, C // 2, 2)
     f0, f1 = tl.split(freq)
     if is_fwd:
